Fix uvx support by renaming py-scripts to py_scripts and updating project config
