+++ py_scripts/s2t/env.example
