+++ py_scripts/s2t/record.sh
