+++ py_scripts/s2t/transcribe.py
