+++ py_scripts/webp_converter/Makefile
